# swagger.json
{
    "swagger": "2.0",
    "info": {
        "version": "1.0.0",
        "title": "PRV Backend",
        "description": "Node JS Backend V0.1",
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        }
    },
    "host": "localhost:5000/",
    "basePath": "",
...
    "schemes": [
        "http"
    ],
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "securityDefinitions": {
        "Bearer": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
    "paths": {
...
        "/admin/register": {
            "post": {
                "tags": [
                    "Admin"
                ],
                "description": "Create new Admin in system",
                "parameters": [{
                    "name": "admin",
                    "in": "body",
                    "description": "Admin that we want to create",
                    "schema": {
                        "$ref": "#/definitions/register"
                    }
                }],
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "New Admin is created",
                        "schema": {
                            "$ref": "#/definitions/register"
                        }
                    }
                }
            }
        },
...
        "/admin/login": {
            "post": {
                "tags": [
                    "Admin"
                ],
                "description": "Create new admin in system",
                "parameters": [{
                    "name": "admin",
                    "in": "body",
                    "description": "Admin that we want to create",
                    "schema": {
                        "$ref": "#/definitions/login"
                    }
                }],
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "New admin is created",
                        "schema": {
                            "$ref": "#/definitions/login"
                        }
                    }
                }
            }
        },
...
        "/admin/getAllAdmins": {
            "get": {
                "security": [{ "Bearer": [] }],
                "tags": [
                    "Admin"
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/User"
                        }
                    }
                },
                "produces": [
                    "application/json"
                ]
            }
        },
        "/admin/getAllUsers": {
            "get": {
                "security": [{ "Bearer": [] }],
                "tags": [
                    "Admin"
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/User"
                        }
                    }
                },
                "produces": [
                    "application/json"
                ]
            }
        },
...
        "/admin/getOne/{email}": {
            "parameters": [{
                "name": "email",
                "in": "path",
                "required": true,
                "description": "Email of admin that we want to find",
                "type": "string"
            }],
            "get": {
                "security": [{ "Bearer": [] }],
                "tags": [
                    "Admin"
                ],
                "responses": {
                    "200": {
                        "description": "Admin is found",
                        "schema": {
                            "$ref": "#/definitions/User"
                        }
                    }
                }
            }
        },
...
        "/superAdmin/login": {
            "post": {
                "tags": [
                    "SuperAdmin"
                ],
                "description": "Login",
                "parameters": [{
                    "name": "superAdmin",
                    "in": "body",
                    "description": "SuperAdmin want to login",
                    "schema": {
                        "$ref": "#/definitions/login"
                    }
                }],
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "Login Successfull",
                        "schema": {
                            "$ref": "#/definitions/login"
                        }
                    }
                }
            }
        },
        "/superAdmin/getOne/{email}": {
            "parameters": [{
                "name": "email",
                "in": "path",
                "required": true,
                "description": "Email of admin that we want to find",
                "type": "string"
            }],
            "get": {
                "security": [{ "Bearer": [] }],
                "tags": [
                    "SuperAdmin"
                ],
                "responses": {
                    "200": {
                        "description": "SuperAdmin is found",
                        "schema": {
                            "$ref": "#/definitions/User"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "login": {
            "required": [
                "email",
                "password"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
...
        "register": {
            "required": [
                "email",
                "password",
                "OTP"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "OTP": {
                    "type": "number"
                }
            }
        },
...
        "User": {
            "required": [
                "id",
                "first_name",
                "last_name",
                "email",
                "card_email",
                "card_img",
                "email_verified_at",
                "dob",
                "nationality",
                "passport_id",
                "maiden_name",
                "pep",
                "contactNumber",
                "password",
                "status",
                "remember_token",
                "created_at",
                "updated_at",
                "countryCode",
                "cardNumber",
                "jdbCardNumber1",
                "jdbCardNumber2",
                "cardStatus",
                "jdbCardStatus1",
                "jdbCardStatus2",
                "google2fa_enable",
                "google2fa_secret",
                "google2fa_qr",
                "countryName",
                "cardCount",
                "appid",
                "is_kyc_approved",
                "kycres",
                "balance_usd",
                "balance_btc",
                "balance_eth",
                "address1",
                "address2",
                "city",
                "state",
                "pincode",
                "d_country",
                "inspectionId",
                "deleted_at",
                "urnNumber",
                "jdbAccountNo1",
                "jdbAccountNo2",
                "is_kyc_details",
                "vendor_id",
                "img_sign",
                "partner_fee",
                "company_name",
                "card_applied",
                "is_doc_downloaded",
                "kyc_country",
                "response_check",
                "passport_file_signature",
                "passport_file_signature_biopic",
                "suspended",
                "suspended_reason",
                "affiliate",
                "emergencycountryCode"
            ],
            "properties": {
                "id": {
                    "type": "string"
                },
                "first_name": {
                    "type": "string"
                },
                "last_name": {
                    "type": "string"
                },
                "userAddress": {
                    "type": "string"
                },
                "email_verified_at": {
                    "type": "string"
                },
                "dob": {
                    "type": "string"
                },
                "nationality": {
                    "type": "string"
                },
                "passport_id": {
                    "type": "string"
                },
                "maiden_name": {
                    "type": "string"
                },
                "pep": {
                    "type": "string"
                },
                "contactNumber": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "status": {
                    "type": "array",
                    "item": "['pending', 'active', 'deactive']"
                },
                "remember_token": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "countryCode": {
                    "type": "integer"
                },
                "cardNumber": {
                    "type": "string"
                },
                "jdbCardNumber1": {
                    "type": "string"
                },
                "jdbCardNumber2": {
                    "type": "string"
                },
                "cardStatus": {
                    "type": "array",
                    "item": "['unpaid','paid','pending','assigned','verified','blocked']"
                },
                "jdbCardStatus1": {
                    "type": "array",
                    "item": "['unpaid','paid','pending','assigned','verified','blocked']"
                },
                "jdbCardStatus2": {
                    "type": "array",
                    "item": "['unpaid','paid','pending','assigned','verified','blocked']"
                },
                "google2fa_enable": {
                    "type": "integer"
                },
                "google2fa_secret": {
                    "type": "string"
                },
                "google2fa_qr": {
                    "type": "string"
                },
                "countryName": {
                    "type": "string"
                },
                "cardCount": {
                    "type": "integer"
                },
                "appid": {
                    "type": "string"
                },
                "is_kyc_approved": {
                    "type": "array",
                    "item": "['0','1']"
                },
                "kycres": {
                    "type": "string"
                },
                "balance_usd": {
                    "type": "string"
                },
                "balance_btc": {
                    "type": "string"
                },
                "balance_eth": {
                    "type": "string"
                },
                "address1": {
                    "type": "string"
                },
                "address2": {
                    "type": "string"
                },
                "city": {
                    "type": "string"
                },
                "state": {
                    "type": "string"
                },
                "pincode": {
                    "type": "string"
                },
                "d_country": {
                    "type": "string"
                },
                "inspectionId": {
                    "type": "string"
                },
                "deleted_at": {
                    "type": "string"
                },
                "urnNumber": {
                    "type": "string"
                },
                "jdbAccountNo1": {
                    "type": "string"
                },
                "jdbAccountNo2": {
                    "type": "string"
                },
                "is_kyc_details": {
                    "type": "array",
                    "item": "['0','1']"
                },
                "vendor_id": {
                    "type": "integer"
                },
                "img_sign": {
                    "type": "string"
                },
                "partner_fee": {
                    "type": "string"
                },
                "company_name": {
                    "type": "string"
                },
                "card_applied": {
                    "type": "integer"
                },
                "is_doc_downloaded": {
                    "type": "integer"
                },
                "kyc_country": {
                    "type": "string"
                },
                "response_check": {
                    "type": "string"
                },
                "passport_file_signature": {
                    "type": "string"
                },
                "passport_file_signature_biopic": {
                    "type": "string"
                },
                "suspended": {
                    "type": "number"
                },
                "suspended_reason": {
                    "type": "string"
                },
                "card_email": {
                    "type": "string"
                },
                "card_img": {
                    "type": "string"
                },
                "affiliate":{
                    "type": "boolean"
                },
                "emergencycountryCode":{
                    "type":"integer"
                }
            }
        },
...
    }
}